+++ backend/metagrid/api_proxy/tests/test_views.py
@@ -79,18 +79,6 @@ class TestProxyViewSet(APITestCase):
         response = self.client.get(url)
         self.assertEqual(response.status_code, status.HTTP_302_FOUND)
 
-    # @responses.activate
-    # def test_wget(self):
-    #     url = reverse("do-wget")
-    #     responses.get(settings.WGET_URL)
-    #     response = self.client.get(
-    #         url,
-    #         {
-    #             "dataset_id": "CMIP6.CMIP.IPSL.IPSL-CM6A-LR.abrupt-4xCO2.r12i1p1f1.Amon.n2oglobal.gr.v20191003|esgf-data1.llnl.gov"
-    #         },
-    #     )
-    #     assert response.status_code == status.HTTP_200_OK
-
     @responses.activate
     def test_search(self):
         url = reverse("do-search")
