+++ backend/metagrid/wget/templates/wget-simple-template.sh
