+++ backend/metagrid/wget/templates/wget-template.sh
@@ -6,12 +6,12 @@
 # Generated by {{request.get_host}} - {{timestamp}}
 # Search URL: {{request.build_absolute_uri}}
 # Request method: {{request.method}}
-# 
+#
 ###############################################################################
 # first be sure it's bash... anything out of bash or sh will break
 # and the test will assure we are not using sh instead of bash
 if [ $BASH ] && [ `basename $BASH` != bash ]; then
-    echo "######## This is a bash script! ##############" 
+    echo "######## This is a bash script! ##############"
     echo "Change the execution bit 'chmod u+x $0' or start with 'bash $0' instead of sh."
     echo "Trying to recover automatically..."
     sleep 1
@@ -178,7 +178,7 @@ if ((update)); then
         mv $0 $0.old.$counter
         echo "$new_wget" > $0
     fi
-    exit 0      
+    exit 0
 fi
 
 check_chksum() {
@@ -240,7 +240,7 @@ remove_from_cache() {
 
 download() {
     wget="wget ${insecure:+--no-check-certificate} ${quiet:+-q} ${quiet:--v}"
-    
+
     while read line
     do
         # read csv here document into proper variables
@@ -251,14 +251,14 @@ download() {
 
         #get the cached entry if any.
         cached="$(grep -e "^$file" "$CACHE_FILE")"
-        
+
         #if we have the cache entry but no file, clean it.
         if [[ ! -f $file && "$cached" ]]; then
             #the file was removed, clean the cache
             remove_from_cache "$file"
             unset cached
         fi
-        
+
         #check it wasn't modified
         if [[ -n "$cached" && "$(get_mod_time_ $file)" == $(echo "$cached" | cut -d ' ' -f2) ]]; then
                     if [[ "$chksum" == "$(echo "$cached" | cut -d ' ' -f3)" ]]; then
@@ -270,7 +270,7 @@ download() {
                 remove_from_cache "$file"
                 unset cached
             else
-                #file on server is different from what we have. 
+                #file on server is different from what we have.
                 echo "WARNING: The remote file was changed (probably a new version is available). Use -U to Update/overwrite"
                 continue
             fi
@@ -286,7 +286,7 @@ download() {
                 #No status will be stored
                 break
             else
-                $wget -O "$file" $url || { failed=1; break; }  
+                $wget -O "$file" $url || { failed=1; break; }
             fi
 
             #check if file is there
@@ -321,7 +321,7 @@ download() {
                                 sleep 1
                                 break
                             fi
-                        
+
                             rm $file
                             #try again
                             echo -n "  re-trying..."
@@ -337,12 +337,12 @@ download() {
             #done!
             break
         done
-        
+
         if ((failed)); then
             echo "download failed"
             unset failed
         fi
-        
+
     done <<<"$download_files"
 
 }
